requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
croniter==3.0.3
discord==2.3.2
discord-py-interactions==5.13.1
discord-typings==0.9.0
discord.py==2.4.0
dnspython==2.6.1
emoji==2.12.1
exceptiongroup==1.2.2
filelock==3.15.4
Flask==3.0.3
frozenlist==1.4.1
gdown==5.2.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
pdf2image==1.17.0
pillow==10.4.0
pymongo==4.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tomli==2.0.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
yarl==1.9.4
